Moved register and un-register callbacks to streaming_services
