+++ src/ansys/fluent/core/streaming_services/events_manager_service.py
@@ -1,5 +1,4 @@
 """Module for events management."""
-from functools import partial
 import itertools
 from typing import Callable, List
 
@@ -31,7 +30,6 @@ class EventsManager(StreamingService):
             streaming_service=service,
         )
         self._session_id: str = session_id
-        self._events_to_callbacks_map: dict = {}
         self._id_iter = itertools.count()
         self._events_list: List[str] = [
             attr for attr in dir(EventsProtoModule) if attr.endswith("Event")
@@ -45,7 +43,7 @@ class EventsManager(StreamingService):
                 event_name = response.WhichOneof("as")
                 with self._lock:
                     self._streaming = True
-                    callbacks_map = self._events_to_callbacks_map.get(event_name, {})
+                    callbacks_map = self._service_callbacks.get(event_name, {})
                     for call_back in callbacks_map.values():
                         call_back(
                             session_id=self._session_id,
@@ -72,49 +70,8 @@ class EventsManager(StreamingService):
     def register_callback(
         self, event_name: str, call_back: Callable, *args, **kwargs
     ) -> str:
-        """Register the callback.
-
-        Parameters
-        ----------
-        event_name : str
-            Event name to register the callback to.
-
-        call_back : Callable
-            Callback to register.
-
-        Returns
-        -------
-        str
-            Registered callback ID.
-
-        Raises
-        ------
-        RuntimeError
-            If event name is not valid.
-        """
         if not event_name in self.events_list:
             raise RuntimeError(f"{event_name} is not a valid event.")
-        with self._lock:
-            event_name = event_name.lower()
-            id = f"{event_name}-{next(self._id_iter)}"
-            callbacks_map = self._events_to_callbacks_map.get(event_name)
-            if callbacks_map:
-                callbacks_map.update({id: partial(call_back, *args, **kwargs)})
-            else:
-                self._events_to_callbacks_map[event_name] = {
-                    id: partial(call_back, *args, **kwargs)
-                }
-            return id
-
-    def unregister_callback(self, callback_id: str):
-        """Unregister the callback.
-
-        Parameters
-        ----------
-        callback_id : str
-            ID of the registered callback.
-        """
-        with self._lock:
-            for callbacks_map in self._events_to_callbacks_map.values():
-                if callback_id in callbacks_map:
-                    del callbacks_map[callback_id]
+        return super().register_callback(
+            service_name=event_name, call_back=call_back, *args, **kwargs
+        )
